SE-91 Fix bug on L187 - InvalidPlacement assertion
The solution:
Added a comparison check before extraction.

The problem:
We were getting an arithmetic error
while asserting that a card placement is valid.

We were doing a *column_card - DIAMONDS_INDEX
which would return a negative number when the
column_card was a HEART card. The the comparison
between a u64 and a negative number would throw an
exception. Same could happen if we did the same
with a CLUBS card minus the SPADES_INDEX.