SE-91 Add valid from_deck_to_pile & fix init_easy_game bug
Bug:
 On easy game initialization, 3 out of 4 piles
 did not contain an ace. Only the first pile did.

Creating a pile by <CLASS>_INDEX and doing
vector::remove was making the following pile
creations obsolete, since the vector was resized
the <CLASS>_INDEX of the other classes was
pointing to the incorrect place.

Fix by subtracting by each class index,
the number of piles created before it (for the other
classes).