SE-108 Fix bug - not able to drop queen on king - column of K
When a column consisting of only one King was to
receive a Queen it failed. That was due to the fact
that the underlying "empty-droppable" component was
overtaking the droppable area of the card on top of
it (the King).

To fix it we use the same check that is used in a normal
column-to-column move, and if that fails,
then we check if the card is King.