SE-165 Separate getGameObjectDetails by digest & by objectId
When we create a new game, sponsoring a transaction
we only have the digest, so we make more calls to
the network to get the game object's contents.

On the other hand, when we need to handle a
transaction failure we already have the game objectId
and thus we can get the contents by directly using
getObject.