SE-165 Sponsor all moves
WARNING: Transactions are way slower

We make more requests because sponsored
transactions return only the digest.

Reasons:
- To get the effects, events etc we need to
get the transaction block details. (1 more request)

- If we also need to get a game's details,
we need to do a getObject request (+1 request)
which depends on the objectId that we get from
the getTransactionBlock request (+1 request).

In total, 2 more requests are needed.