SE-165 Fix bug on failure of 'end-game' button
If the end-game (i.e. delete unfinished game) transaction
failed, the game would not be re-rendered in
the correct state.

The game would show up like being in its initial
state even though some moves have been made.

To fix it, I replaced the setSpinning screen with
a spinning cursor. This way the gameboard does not
have to be re-rendered and be prone to inconsistencies.